A working commit, that implements most of the building blocks required for the model covariance.
In 2D, Weerachai's model covariance is used. It has been renamed to Diffusion (since he is solving the diffusion eqn), and stored in 2D_MT/modelParam/modelCov/. It works.

In 3D, currently implementing our recursive autoregression scheme (RecursiveAR, stored in 3D_MT/modelParam/modelCov/). Neither that nor Weerachai's covariance are not yet ready to use, but the respective model covariance structures already exist, along with most of the other pieces. In particular, I've added an iscalar type to implement the mask that defines scaling and smoothing surfaces. In the process of doing so, I've added adequate read and write routines for all scalar types in EMfield/sg_scalar (not all of them tested yet). Moreover, a covariance input file can now be specified and read, to initialize the covariance. I'll provide an example in the next commit.

Other than the 3D covariance, everything (including 3D NLCG) compiles and runs.