1) Corrected an obvious bug in create_CmSqrt for recursive autoregression.    This bug caused the program to stop if run without a model covariance.    NOTE: currently applying some sort of covariance by default; probably    not a good idea. 2) Added a write_Cond3D statement to NLCG. 3) Updated the screen message in UserCtrl.
So you see, no significant changes here. Unfortunately, the NLCG program
DOES NOT RUN right now on the examples that I tried. It bombs with a
segmentation fault even if I comment out the covariance (and replace
it with m_out = m_in). It is unhappy about something in initSolver
the first time it tries to compute the gradient. This is an unpleasant
surprise: I seem to remember that everything at least run when I stopped
working on this couple of weeks ago. Will try to debug this asap.