Global spherical code updated to match the modular system more closely. The model parameter now includes the crust (modeldef), and the operators have been renamed and are ready to use (model_operators).
Fixed a bug that was there for a while but was never a problem since
I've only used the responses & derivative computations lately. This came
up in calc_symmetry (i.e. Jmult) computations... The layer pointer 
in vectorPj (jacobian) was never set, resulting in a segmentation
fault. Note: takes too much memory with a degree 9 parametrization
to run this on MT!!! DON'T DO THAT!!! 

The path of minimum resistance is to set up its own upper level modules
for the time being (DataSpace and SensMatrix). An immediate goal is to
start using the modular system inversion algorithms. Also, to continue
development in the modular framework, e.g. be specific about the EMsoln
business (currently called SolnRHS). Should make it easier once we start 
working on the boundary conditions.

Some of this is already written, but hasn't been used or debugged yet.