Ok, so the global NLCG works. 2D AND 3D MT NOT TESTED YET!
I don't think I've broken something, but please do not update until I test.

Bugs in the global code that need to be debugged asap:

1. I've lost a sign somewhere in the gradient computations. NLCG didn't
converge because the gradient was negative of what it should be. This
error (just like the other bugs) is inherited from the old version of the
code. Previously I dealt with this by multiplying the gradient by -2
instead of +2. Here, I also inserted an extra minus sign in JmultT to 
get NLCG (and the derivative test) to work.

   When fixing this, please note that all the symmetry tests are passed
at the moment; so the error must clearly be duplicated in both the adjoint
and the direct computations.

2. The weird behaviour at the poles when the resolution is increased.

3. The boundary conditions should not include the vertical components.

4. A memory leak in NLCG: it seems to require progressively more memory
when it runs. This problem may make it unpractical to run full-size
inversions, both for 3D MT and the global problem. The memory leak may
also be in the forward solver and/or gradient computations in both codes.

5. Sometimes the NLCG doesn't exit gracefully. This happens when the search
step alpha is too close to zero, but the sufficient decrease condition for
some reason isn't satisfied. The models and gradients become crazy and we
end up with NaNs for f, RMS and alpha. Need to track down why and when
that happens and to make sure it exits nicely in this case.


Features in the global code that need to be added soon:

6. write_modelParam (doesn't yet exist, instead prints info to screen)

7. write_data & read_data (need to be generic - we want to use them in NLCG!)

8. write_sensMatrix & read_sensMatrix

9. calcSensMatrix & Jmult (don't exist as separate routines for the global code).