Used MemoryScape to find the remaining memory leaks. No leaks remaining! Figured out how to deal with the temporary arrays created by functions. It is now completely clear where the leaks come from. Wrote documentation. If certain rules are observed in any new code, as described in the documentation (avoiding_memleak.pdf) we shouldn't encounter memory leaks again. Anyone working on this code should read this before writing basic modules!!!
2D and 3D MT are now clean of memory leaks (except for a minor bug
in the input subroutine for 3D data, where the compids are allocated
for every frequency... not fixed since this subroutine will be
rewritten to use the dictionaries anyway). Global code is also clean
of memory leaks, but requires a lot of rewriting. Check for any newly
appearing memory leaks once in a while! 
The inversion runs, but need to fiddle with the options and synthetic tests 
to ensure convergence for the global problem. Currently, it stalls by eventually 
making alpha too small... never exits the cubic line search.

Next: 1) debug the poles problem and the boundary values for the global;
2) play with the settings for the global and 3D MT to ensure convergence
on realistic problems (set up a configuration file for the inverse solver!);
3) then, finish setting up the model and data output for the inversion -
write_data subroutines for MT and write_model for the global code.
Then use the code!