IMPORTANT BUG FIX
In the past revisions, ALL data were converted to SI units using a factor
computed from the information in the file. However, vertical transfer functions
are of course dimensionless and should NOT be affected by any units conversions.
Therefore, in the previous revisions the inversion was working correctly only if
the data were originally in the SI units [V/m]/[T], or if the vertical field
transfer functions were not used. This has now been corrected.
Thanks to Naser Meqbel for finding the bug.

Also, added the option of using the vertical field transfer functions alone
(Gary edited the DataFunc module). To the best of my knowledge, nothing has been
broken. The new capability should work, but has not been tested.