FIXED A MAJOR BUG in create/copy EMsoln, that caused the NLCG in revisions after update 199 incl. to NOT WORK. Rather, it would run, but never converge, since the transmitters were all mixed up... the bug was completely mysterious, though. We should carefully deallocate everything in the create routine, if there is any chance that any of the input parameters have changed between calls. Also added some interface statements and some additional cleaning to the other create/copy routines in SolnRHS, to avoid these problems in the future.
Temporarily reverted the changes in the EMsolver / 3D FWD related to saving the model parameter for boundary
conditions computations. Will need to spend some time and figure out why the standard NLCG tests don't converge
when these changes are applied!

This commit should be fully functional again. Works with the legacy data format (same as in revisions 100-198).
The new list format for the data has not been completed and debugged yet. 

Both 2D and 3D MT NLCG inversions are tested and work in this commit.