Finally figured out why the global code compiled with Intel compiler produced segmentation faults on 2x2 degree or finer grids. At the same time, the code was working fine on coarser grids... and with PGI.
Turns out the segmentation fault occurs at the first rscalar linear
combination (line 120 in ForwardSolver.f90). The reason is that there
isn't enough stack storage to perform the addition of 2x2 degree arrays.
All allocatable arrays are always stored in heap, but the temporary
automatic arrays that are created during the summation are stored in stack.
The stack memory is exceeded, causing a silent segmentation fault with
no traceback.

Solutions are described here:
http://software.intel.com/en-us/articles/intel-fortran-compiler-increased-stack-usage-of-80-or-higher-compilers-causes-segmentation-fault/

The user has to increase the stack memory on the machine beforehand:

ulimit -s unlimited     for bash shell, or
limit stacksize unlimited     for csh shell

OR use the -heap-arrays [n] compiler option that stores any array
bigger than n kb in heap rather than stack. However, this is known
(for some reason) to be incompatible with OpenMP.
When the -heap-arrays option is specified, the -parallel option is
ignored, and the code runs in serial without the use of OpenMP.

For now, I am ignoring this incompatibility (the code still runs)
and adding the -heap-arrays option to the default MakefileG.
This is because I don't want to spend another full day debugging this,
when I encounter this problem again on a different machine.
However, I strongly recommend increasing the stack size instead
if you are to use this code repeatedly on fine sized grids.
This will result in greater efficiency: my 2x2 degree example
just run for 58.21 mins with -heap-arrays, vs 40.03 mins with
unlimited stack size (and without this compiler option).

Strange, but for some reason the global code only uses one CPU
for the forward computations, even though I compile it with -parallel
compiler and linker options. Need to check what is going on... it
might be worthwhile to quickly parallelize some loops with OpenMP.