Change in NLCG output: the residual will no longer be normalized by errors.
Also, for the global code, I changed the way the model output behaves.
We output both the final "smooth" model m = Cm^{1/2} mhat + m0,
and the "rough" model mhat, which is the result of the inverse search.
The smooth model m is best used to save and to plot the model; while
the rough model mhat is best used in conjunction with the prior to
restart the inversion, or compute the responses / derivatives.
Therefore, to avoid confusion we DO NOT write the regularization
parameters for a model if the smoothing has already been applied.
Regularization parameters are only needed to smooth a rough model,
so we leave them out for m (and leave them in for mhat).