Important change in MPI_Main. I now make sure that the solution computed by fwdPred_TX (e0) is not initialized in MPI_Main. We leave it to initSolver to initialize this. Then, we can always use the previous solution as the starting solution, if we choose to do so in initSolver routine.
To facilitate copying the multi-transmitter EM solution (eAll),
I now allow copying of non-allocated solution vectors in
copy_solnVectorMTX. I think this is more reasonable anyway.
There is no reason why they need to be allocated in order to
copy the multi-transmitter structure.

MPI problems solved in the global code:
1. all processors pointed to the global grid which was initialized
and stored on the master node, so grid wasn't accessible to the
forward solver;
2. global variables such as the array dimensions in InitGlobalArrays
were not initialized on the worker nodes. Worker nodes call the 
subroutine InitGlobalData to initialize everything they will need.
Including InitGlobalArrays inside of InitGlobalData solved this;
3. cUserDef is a global structure that is needed by the ForwardSolver
module even after the global data are initialized. This has to be
initialized as a global variable inside the InitGlobalData routine.

The global code now works with the MPI, both for the forward
modelling and the inversion. Naser will test the 3D MT, but
there is no reason why it should be broken - so hopefully this
will also work.