Revert the changes in Main.f90 files: Don't get grid and AllData as outputs from initGlobalData
In the new main program files (2D-MT,3D-MT and Global3D):
After initializing the cluster, only the Master reads the command line arguments and creates sUserDef. Afterwards, the Master distributes sUserDef to all Workers. Before doing anything, the Master (as before) as well as the worker initialize the Global Data by calling initGlobalData in the main program.
Because now, we are not using SensComp.f90 and main.f90 in MPI_Main anymore, a new subroutine to set the grid in MPI_Main is required.