Updating to trunk revision 316: new features and bug fixes.
1. The computation of the full J-matrix is now implemented in parallel. 
The parallelization is over periods, stations and data types. 
An optimal number of processors would be nTx* nStn *nDt.

Right now, the worker processors send blocks of J-matrix back to the master processors 
who sort out the blocks in the J-matrix and write it out in a file. This is done ONLY for test 
proposes.
In the near future, the worker will keep the blocks of J-matrix on their own memory. 
Any further mathematical operation with J-matrix (i.e. JJT, JTJ or JTd) will be done in parallel.

2. Added the inverse of the covariance operator, which may be applied using the -C option. 
The covariance is NOT YET RESCALED in this commit and works identically to the previous version. 

Usage: -C  [FWD|INV] rFile_Model wFile_Model [rFile_Cov rFile_Prior]

 -C FWD rFile_Model wFile_Model [rFile_Cov rFile_Prior]
  Applies the model covariance to produce a smooth model output
  Optionally, also specify the prior model to compute resistivities
  from model perturbation: m = C_m^{1/2} \tilde{m} + m_0

 -C INV rFile_Model wFile_Model [rFile_Cov rFile_Prior]
  Applies the inverse of model covariance (if implemented)
  Optionally, also specify the prior model to compute starting
  perturbation for the inversion: \tilde{m} = C_m^{-1/2} (m - m_0)
  WARNING: may be poorly conditioned if the smoothing is too strong;
  always check your model for white noise after using this option.

IMPORTANT:
The covariance is a poorly conditioned operator, essentially multiplying the
small scale features by zero; its inverse is therefore not well defined.
For example, if you apply alpha=0.1 covariance over 8 times, or alpha=0.3
covariance over 3 times, multiplication by the inverse covariance will produce
white noise of a large magnitude (3-4 orders of magnitude, comparable to that of
our models). It is safe to apply a strong covariance, but not to apply its inverse.
Until we find a way to regularize this, I am not making this part of the inversion,
but rather a separate option that may be used with care.

3. Fixed the 2D MT MPI.