Updating to trunk revision 327:
2D MT BUG FIX.

In 2D MT, the thickness of the air layers depends on the top Earth layer
in the model. The first air layer matches that, the next one gets multiplied
by the factor of 3, and so on for 10 layers total. Thus, the total thickness
of the air layers is proportional to the thickness of the top Earth layer.
This was a serious problem since the vertical site location in the data file
had to be defined using distance from the top of the air domain. Z coordinate
was therefore dependent on the thickness of the air layers. This is now fixed:
zero Z coordinate in the data file corresponds to the Earth's surface;
positive is down; negative is up. If we had topography, Earth's surface would
be defined by the highest point in the topography and some site Z coordinates
would be positive to reflect the topography.
Also fixed the Z coordinate in examples and Matlab scripts.


3D MT NEW FEATURE.

In 3D MT, added the code by Kristina Tietze (GFZ) that implements inversion
of log apparent resistivity and phases and of the phase tensor. This has been
somewhat tested but not thoroughly debugged. Her 3D MT example called ObliqueOne
is also added.

Inverse configuration scripts in 3D examples have also been updated. Not all
examples are tested to converge with these parameters; e.g. DSM1 doesn't seem
to converge: it stalls at rms of around 8.9. I don't remember whether it converged
before, I think so but I am not sure. In either case, need to figure out why,
but I don't have the time to spend on this. Perhaps later.