Updating to trunk revision 373:
Stability update before releasing the code to the academic community.
Includes updates to the user guide, and the paper currently in press in GJI.
Also, fixes 2D MT inversion in MPI; changes the behaviour of apparent
resistivity and phase inversion (linear, NOT log apparent resistivities
expected in the input file); also includes minor corrections to the
Matlab I/O scripts and some cosmetic changes to the makefiles.