Several very substantial updates that should not affect functionality but have to do with internal structure and the I/O.
1) The internal dictionaries (transmitter type, transmitter, receiver and data type) are now fully up-to-date
   to allow for joint EM modeling and inversion. This only creates the background structure; the rest of the joint
   functionality has not been added yet.

2) The data file has now been updated to the "new" (as of 2018) format. Specifically, there is now an optional additional
   third line in the header (directly following the two "#" data block header lines) that starts with a "+" sign
   and is used to specify that transmitter type. For traditional MT purposes, it will look like "+ MT".
   The data I/O is fully backwards compatible: if the additional line is not found, it assumes MT and writes old style.

3) The output for the data sensitivity matrix is substantially updated for usability, as in Gary's comment below.

I did just make changes to SensitivityMatrix because I am trying to use sensitivity matrix computed by ModEM in ModEMMinv matlab codes, for work with Williams. There were a few 
issues with what was implemented, that you should add to your version:  (1) Not enough metadata was in sensitivity file for the J that was output to be actually usable -- tx, rx 
are written, but no TxDict and RxDict, which are needed to actually interpret/use!   This is fixed now.   Dependencies change (SensMatrix needs to use the dictionary modules).   
(2)  the sensitivity matrix multiplication did not know about dataType -- this is in the data vector, which is output, but not input.   So data type was left set to 0, and then 
nothing was written to the output file for J*dm.   I added iDt property to SensVector (might not be remembering correct dataType names), so that this is available when doing the 
multiply, and hence set correctly in the D*dm output file.   Those fixes are in the gitlab working version.  

Because the dependencies have changed again, makefiles will need to be remade. But no hurry - the upcoming update will
require them to be updated again. So might as well wait a little.